Move HikariCPJdbcDataSource into a separate module
