+++ slick-hikaricp/src/main/scala/slick/jdbc/hikaricp/HikariCPJdbcDataSource.scala
@@ -1,0 +1,58 @@
+package slick.jdbc.hikaricp
+
+import java.sql.{Driver, Connection}
+import com.typesafe.config.Config
+import slick.SlickException
+import slick.jdbc.{JdbcDataSourceFactory, JdbcDataSource}
+import slick.util.ConfigExtensionMethods._
+
+/** A JdbcDataSource for a HikariCP connection pool.
+  * See `slick.jdbc.JdbcBackend#Database.forConfig` for documentation on the config parameters. */
+class HikariCPJdbcDataSource(val ds: com.zaxxer.hikari.HikariDataSource, val hconf: com.zaxxer.hikari.HikariConfig) extends JdbcDataSource {
+  def createConnection(): Connection = ds.getConnection()
+  def close(): Unit = ds.close()
+}
+
+object HikariCPJdbcDataSource extends JdbcDataSourceFactory {
+  import com.zaxxer.hikari._
+
+  def forConfig(c: Config, driver: Driver, name: String, classLoader: ClassLoader): HikariCPJdbcDataSource = {
+    if(driver ne null)
+      throw new SlickException("An explicit Driver object is not supported by HikariCPJdbcDataSource")
+    val hconf = new HikariConfig()
+
+    // Connection settings
+    hconf.setDataSourceClassName(c.getStringOr("dataSourceClass", null))
+    Option(c.getStringOr("driverClassName", c.getStringOr("driver"))).map(hconf.setDriverClassName _)
+    hconf.setJdbcUrl(c.getStringOr("url", null))
+    c.getStringOpt("user").foreach(hconf.setUsername)
+    c.getStringOpt("password").foreach(hconf.setPassword)
+    c.getPropertiesOpt("properties").foreach(hconf.setDataSourceProperties)
+
+    // Pool configuration
+    hconf.setConnectionTimeout(c.getMillisecondsOr("connectionTimeout", 1000))
+    hconf.setValidationTimeout(c.getMillisecondsOr("validationTimeout", 1000))
+    hconf.setIdleTimeout(c.getMillisecondsOr("idleTimeout", 600000))
+    hconf.setMaxLifetime(c.getMillisecondsOr("maxLifetime", 1800000))
+    hconf.setLeakDetectionThreshold(c.getMillisecondsOr("leakDetectionThreshold", 0))
+    hconf.setInitializationFailFast(c.getBooleanOr("initializationFailFast", false))
+    c.getStringOpt("connectionTestQuery").foreach { s =>
+      hconf.setJdbc4ConnectionTest(false)
+      hconf.setConnectionTestQuery(s)
+    }
+    c.getStringOpt("connectionInitSql").foreach(hconf.setConnectionInitSql)
+    val numThreads = c.getIntOr("numThreads", 20)
+    hconf.setMaximumPoolSize(c.getIntOr("maxConnections", numThreads * 5))
+    hconf.setMinimumIdle(c.getIntOr("minConnections", numThreads))
+    hconf.setPoolName(name)
+    hconf.setRegisterMbeans(c.getBooleanOr("registerMbeans", false))
+
+    // Equivalent of ConnectionPreparer
+    hconf.setReadOnly(c.getBooleanOr("readOnly", false))
+    c.getStringOpt("isolation").map("TRANSACTION_" + _).foreach(hconf.setTransactionIsolation)
+    hconf.setCatalog(c.getStringOr("catalog", null))
+
+    val ds = new HikariDataSource(hconf)
+    new HikariCPJdbcDataSource(ds, hconf)
+  }
+}
